+++ Base/Devices/imstkVRPNDeviceServer.h
@@ -74,7 +74,7 @@ public:
     ///
     /// \brief Add device
     ///
-    void addDevice(std::string deviceName, DeviceType deviceType, size_t id = 0);
+    void addDevice(std::string deviceName, DeviceType deviceType, int id = 0);
 
 protected:
 
@@ -98,9 +98,9 @@ private:
     const std::string m_machine; ///< machine name or IP
     const int m_port;            ///< connection port
 
-    std::map<std::string, std::pair<DeviceType, size_t>> m_deviceInfoMap;///< list of iMSTK client info
-    vrpn_Connection * m_serverConnection;                                ///< VRPN server connection
-    vrpn_MainloopContainer* m_deviceConnections;                         ///< VRPN device connections
+    std::map<std::string, std::pair<DeviceType, int>> m_deviceInfoMap;///< list of iMSTK client info
+    vrpn_Connection * m_serverConnection;                             ///< VRPN server connection
+    vrpn_MainloopContainer* m_deviceConnections;                      ///< VRPN device connections
 
 };
